.scripts/requirements.txt
functions-framework==3.5.0
google-api-python-client==2.112.0
google-generativeai==0.6.0
google-auth==2.27.0